ai-service/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
numpy
scipy
nltk
opencv-python-headless
mediapipe
